refactor(parser): expose primary weapon and split lib.rs into modules
Adds PlayerState.primary (the player's stowed main gun) to the WASM
parser, derived from the pawn inventory (m_hMyWeapons): a primary gun
wins outright, else the first pistol carried; empty for knife/utility only.

Splits the 2.1k-line lib.rs into focused modules with no behavior change
(byte-identical Replay JSON):
  - schema.rs    output serde types
  - props.rs     entity/game-event/team/pause readers
  - weapons.rs   weapon label tables (const) + inventory helpers
  - collector.rs the streaming observer + Collector state
  - assemble.rs  build_replay + voice blob packing
  - lib.rs       wasm-bindgen surface only (~150 lines)
